+++ src/util/Hosts.cpp
@@ -1,7 +1,7 @@
 /*
  * Hosts.cpp
  *
- * The hosts database
+ * Hosts database
  *
  *
  * Copyright (C) 2018 Amit Kumar (amitkriit@gmail.com)
@@ -12,12 +12,11 @@
 
 #include "Hosts.h"
 #include "../base/common/Exception.h"
-#include "../base/common/Logger.h"
 #include <cstring>
 
 namespace {
-/**
- * Helper functions for formatting a hosts file.
+/*
+ * Helper functions
  */
 static void writeTuple(FILE *f, unsigned long long uid, const char *host,
 		const char *service) noexcept {
@@ -49,26 +48,31 @@ Hosts::Hosts(const char *path, bool readOnly) {
 }
 
 Hosts::~Hosts() {
-	clear();
+	close();
 }
 
 void Hosts::open(const char *path, bool readOnly) {
 	try {
-		clear();
-		openConnection(path, readOnly);
+		auto flags = readOnly ? SQLite::READ_ONLY : SQLite::RW_CREATE;
+		SQLite::open(path, flags);
+		closeStatements();
 		if (!readOnly) {
 			createTable();
 		}
 		prepareStatements();
 	} catch (const BaseException &e) {
-		//Clean up to prevent resource leak
-		clear();
+		close();
 		throw;
 	}
 }
 
-void Hosts::batchUpdate(const char *path) {
-	if (!db.conn || Storage::testFile(path) != 1) {
+void Hosts::close() noexcept {
+	closeStatements();
+	SQLite::close();
+}
+
+void Hosts::load(const char *path) {
+	if (Storage::testFile(path) != 1) {
 		throw Exception(EX_RESOURCE);
 	}
 
@@ -78,7 +82,7 @@ void Hosts::batchUpdate(const char *path) {
 	}
 
 	try {
-		beginTransaction();
+		SQLite::transact(SQLiteStage::BEGIN);
 		//-----------------------------------------------------------------
 		char line[2048];
 		char format[64];
@@ -98,133 +102,128 @@ void Hosts::batchUpdate(const char *path) {
 		}
 		Storage::closeStream(f);
 		//-----------------------------------------------------------------
-		endTransaction();
+		SQLite::transact(SQLiteStage::COMMIT);
 	} catch (const BaseException &e) {
 		Storage::closeStream(f);
 		throw;
 	}
 }
 
-void Hosts::batchDump(const char *path, int version) {
-	if (db.conn) {
-		//-----------------------------------------------------------------
-		auto query =
-				"SELECT uid, name, service, type FROM hosts ORDER BY uid ASC";
-		sqlite3_stmt *stmt = nullptr;
-		if (sqlite3_prepare_v2(db.conn, query, strlen(query), &stmt,
-				nullptr) != SQLITE_OK) {
-			finalize(stmt);
-			throw Exception(EX_STATE);
-		}
-		//-----------------------------------------------------------------
-		auto f = Storage::openStream(path, "w");
-		if (!f) {
-			finalize(stmt);
-			throw Exception(EX_ARGUMENT);
-		}
+void Hosts::dump(const char *path, int version) {
+	auto query = "SELECT uid, name, service, type FROM hosts ORDER BY uid ASC";
+	auto stmt = SQLite::prepare(query);
 
-		writeHeading(f, version);
-		while (sqlite3_step(stmt) == SQLITE_ROW) {
-			auto id = (unsigned long long) sqlite3_column_int64(stmt, 0);
-			auto host = (const char*) sqlite3_column_text(stmt, 1);
-			auto service = (const char*) sqlite3_column_text(stmt, 2);
-			auto type = sqlite3_column_int(stmt, 3);
-			if (version == 1) {
-				writeTuple(f, id, host, service, type);
-			} else {
-				writeTuple(f, id, host, service);
-			}
-		}
-		Storage::closeStream(f);
-		finalize(stmt);
-	} else {
-		throw Exception(EX_RESOURCE);
+	auto f = Storage::openStream(path, "w");
+	if (!f) {
+		SQLite::finalize(stmt);
+		throw Exception(EX_ARGUMENT);
 	}
+
+	writeHeading(f, version);
+	while (SQLite::step(stmt) == SQLITE_ROW) {
+		auto id = (unsigned long long) SQLite::columnLongInteger(stmt, 0);
+		auto host = (const char*) SQLite::columnText(stmt, 1);
+		auto service = (const char*) SQLite::columnText(stmt, 2);
+		auto type = SQLite::columnInteger(stmt, 3);
+		if (version == 1) {
+			writeTuple(f, id, host, service, type);
+		} else {
+			writeTuple(f, id, host, service);
+		}
+	}
+	Storage::closeStream(f);
+	SQLite::finalize(stmt);
 }
 
 int Hosts::get(unsigned long long uid, NameInfo &ni) noexcept {
-	if (db.conn && db.qStmt) {
-		int ret = 0;
-		int z;
-		memset(&ni, 0, sizeof(ni));
-		sqlite3_bind_int64(db.qStmt, 1, uid);
-		if ((z = sqlite3_step(db.qStmt)) == SQLITE_ROW) {
-			strncpy(ni.host, (const char*) sqlite3_column_text(db.qStmt, 0),
-					sizeof(ni.host) - 1);
-			strncpy(ni.service, (const char*) sqlite3_column_text(db.qStmt, 1),
-					sizeof(ni.service) - 1);
-			ni.type = sqlite3_column_int(db.qStmt, 2);
-		} else if (z == SQLITE_DONE) {
-			//No record found
-			ret = 1;
-		} else {
-			ret = -1;
-		}
-		reset(db.qStmt);
-		return ret;
-	} else {
+	if (!stmt.pq) {
 		return -1;
 	}
+
+	int ret = 0;
+	int z;
+	memset(&ni, 0, sizeof(ni));
+	SQLite::bindLongInteger(stmt.pq, 1, uid);
+	if ((z = SQLite::step(stmt.pq)) == SQLITE_ROW) {
+		strncpy(ni.host, (const char*) SQLite::columnText(stmt.pq, 0),
+				sizeof(ni.host) - 1);
+		strncpy(ni.service, (const char*) SQLite::columnText(stmt.pq, 1),
+				sizeof(ni.service) - 1);
+		ni.type = SQLite::columnInteger(stmt.pq, 2);
+	} else if (z == SQLITE_DONE) {
+		//No record found
+		ret = 1;
+	} else {
+		ret = -1;
+	}
+	SQLite::reset(stmt.pq);
+	SQLite::unbind(stmt.pq);
+	return ret;
 }
 
 int Hosts::put(unsigned long long uid, const NameInfo &ni) noexcept {
-	if (db.conn && db.iStmt) {
-		int ret = 0;
-		sqlite3_bind_int64(db.iStmt, 1, uid);
-		sqlite3_bind_text(db.iStmt, 2, ni.host, strlen(ni.host), SQLITE_STATIC);
-		sqlite3_bind_text(db.iStmt, 3, ni.service, strlen(ni.service),
-		SQLITE_STATIC);
-		sqlite3_bind_int(db.iStmt, 4, ni.type);
-		if (sqlite3_step(db.iStmt) != SQLITE_DONE) {
-			ret = -1;
-		}
-		reset(db.iStmt);
-		return ret;
-	} else {
+	if (!stmt.pi) {
 		return -1;
 	}
+
+	int ret = 0;
+	SQLite::bindLongInteger(stmt.pi, 1, uid);
+	SQLite::bindText(stmt.pi, 2, ni.host, strlen(ni.host), SQLiteScope::STATIC);
+	SQLite::bindText(stmt.pi, 3, ni.service, strlen(ni.service),
+			SQLiteScope::STATIC);
+	SQLite::bindInteger(stmt.pi, 4, ni.type);
+	if (SQLite::step(stmt.pi) != SQLITE_DONE) {
+		ret = -1;
+	}
+	SQLite::reset(stmt.pi);
+	SQLite::unbind(stmt.pi);
+	return ret;
 }
 
 int Hosts::remove(unsigned long long uid) noexcept {
-	if (db.conn && db.dStmt) {
-		int ret = 0;
-		sqlite3_bind_int64(db.dStmt, 1, uid);
-		if (sqlite3_step(db.dStmt) != SQLITE_DONE) {
-			ret = -1;
-		}
-		reset(db.dStmt);
-		return ret;
-	} else {
+	if (!stmt.pd) {
 		return -1;
 	}
+
+	int ret = 0;
+	SQLite::bindLongInteger(stmt.pd, 1, uid);
+	if (SQLite::step(stmt.pd) != SQLITE_DONE) {
+		ret = -1;
+	}
+	SQLite::reset(stmt.pd);
+	SQLite::unbind(stmt.pd);
+	return ret;
 }
 
 int Hosts::list(unsigned long long uids[], unsigned int &count,
 		int type) noexcept {
 	if (count == 0) {
 		return 0;
-	} else if (uids && db.conn && db.lStmt) {
-		sqlite3_bind_int(db.lStmt, 1, type);
-		sqlite3_bind_int(db.lStmt, 2, count);
-		unsigned int x = 0;
-		int z = 0;
-		while ((x < count) && ((z = sqlite3_step(db.lStmt)) == SQLITE_ROW)) {
-			uids[x++] = sqlite3_column_int64(db.lStmt, 0);
-		}
-		reset(db.lStmt);
-		count = x;
-		if (z == SQLITE_ROW || z == SQLITE_DONE) {
-			return 0;
-		} else {
-			return -1;
-		}
-	} else {
+	}
+
+	if (!uids || !stmt.pl) {
 		count = 0;
+		return -1;
+	}
+
+	SQLite::bindInteger(stmt.pl, 1, type);
+	SQLite::bindInteger(stmt.pl, 2, count);
+	unsigned int x = 0;
+	int z = 0;
+	while ((x < count) && ((z = SQLite::step(stmt.pl)) == SQLITE_ROW)) {
+		uids[x++] = SQLite::columnLongInteger(stmt.pl, 0);
+	}
+	SQLite::reset(stmt.pl);
+	SQLite::unbind(stmt.pl);
+	count = x;
+	if (z == SQLITE_ROW || z == SQLITE_DONE) {
+		return 0;
+	} else {
 		return -1;
 	}
 }
 
-void Hosts::createDummy(const char *path, int version) {
+void Hosts::dummy(const char *path, int version) {
 	auto f = Storage::openStream(path, "w");
 	if (f) {
 		auto host = "127.0.0.1";
@@ -245,45 +244,13 @@ void Hosts::createDummy(const char *path, int version) {
 	}
 }
 
-void Hosts::clear() noexcept {
-	closeStatements();
-	closeConnection();
-}
-
-void Hosts::openConnection(const char *path, bool readOnly) {
-	closeConnection();
-	int flags = readOnly ?
-	SQLITE_OPEN_READONLY :
-							(SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE);
-	sqlite3 *conn = nullptr;
-	if (sqlite3_open_v2(path, &conn, flags, nullptr) != SQLITE_OK) {
-		WH_LOG_DEBUG("%s", sqlite3_errmsg(conn));
-		//sqlite3_open_v2 always returns a connection handle
-		sqlite3_close(conn);
-		throw Exception(EX_OPERATION);
-	} else {
-		db.conn = conn;
-	}
-}
-
-void Hosts::closeConnection() noexcept {
-	sqlite3_close_v2(db.conn);
-	db.conn = nullptr;
-}
-
 void Hosts::createTable() {
 	auto tq = "CREATE TABLE IF NOT EXISTS hosts ("
 			"uid INTEGER NOT NULL UNIQUE ON CONFLICT REPLACE,"
 			"name TEXT NOT NULL DEFAULT '127.0.0.1',"
 			"service TEXT NOT NULL DEFAULT '9000',"
 			"type INTEGER NOT NULL DEFAULT 0)";
-	if (!db.conn
-			|| sqlite3_exec(db.conn, tq, nullptr, nullptr, nullptr) != SQLITE_OK) {
-		WH_LOG_DEBUG("Could not create database tables");
-		throw Exception(EX_OPERATION);
-	} else {
-		//SUCCESS
-	}
+	SQLite::execute(tq);
 }
 
 void Hosts::prepareStatements() {
@@ -292,77 +259,36 @@ void Hosts::prepareStatements() {
 	auto dq = "DELETE FROM hosts WHERE uid=?";
 	auto lq = "SELECT uid FROM hosts WHERE type=? ORDER BY RANDOM() LIMIT ?";
 
-	closeStatements();
-	if (!db.conn
-			|| (sqlite3_prepare_v2(db.conn, iq, strlen(iq), &db.iStmt, nullptr)
-					!= SQLITE_OK)
-			|| (sqlite3_prepare_v2(db.conn, sq, strlen(sq), &db.qStmt, nullptr)
-					!= SQLITE_OK)
-			|| (sqlite3_prepare_v2(db.conn, dq, strlen(dq), &db.dStmt, nullptr)
-					!= SQLITE_OK)
-			|| (sqlite3_prepare_v2(db.conn, lq, strlen(lq), &db.lStmt, nullptr)
-					!= SQLITE_OK)) {
+	try {
 		closeStatements();
-		WH_LOG_DEBUG("Could not create prepared statements");
-		throw Exception(EX_OPERATION);
-	} else {
+		stmt.pi = SQLite::prepare(iq);
+		stmt.pq = SQLite::prepare(sq);
+		stmt.pd = SQLite::prepare(dq);
+		stmt.pl = SQLite::prepare(lq);
 		resetStatements();
+	} catch (...) {
+		closeStatements();
+		throw Exception(EX_OPERATION);
 	}
 }
 
 void Hosts::resetStatements() noexcept {
-	reset(db.iStmt);
-	reset(db.qStmt);
-	reset(db.dStmt);
-	reset(db.lStmt);
+	SQLite::reset(stmt.pi);
+	SQLite::unbind(stmt.pi);
+	SQLite::reset(stmt.pq);
+	SQLite::unbind(stmt.pq);
+	SQLite::reset(stmt.pd);
+	SQLite::unbind(stmt.pd);
+	SQLite::reset(stmt.pl);
+	SQLite::unbind(stmt.pl);
 }
 
 void Hosts::closeStatements() noexcept {
-	finalize(db.iStmt);
-	finalize(db.qStmt);
-	finalize(db.dStmt);
-	finalize(db.lStmt);
-
-	db.iStmt = nullptr;
-	db.qStmt = nullptr;
-	db.dStmt = nullptr;
-	db.lStmt = nullptr;
-}
-
-void Hosts::reset(sqlite3_stmt *stmt) noexcept {
-	if (stmt) {
-		sqlite3_clear_bindings(stmt);
-		sqlite3_reset(stmt);
-	}
-}
-
-void Hosts::finalize(sqlite3_stmt *stmt) noexcept {
-	reset(stmt);
-	sqlite3_finalize(stmt); //Harmless no-op on nullptr
-}
-
-void Hosts::beginTransaction() {
-	if (!db.conn
-			|| sqlite3_exec(db.conn, "BEGIN", nullptr, nullptr, nullptr)
-					!= SQLITE_OK) {
-		throw Exception(EX_OPERATION);
-	}
-}
-
-void Hosts::endTransaction() {
-	if (!db.conn
-			|| sqlite3_exec(db.conn, "COMMIT", nullptr, nullptr, nullptr)
-					!= SQLITE_OK) {
-		throw Exception(EX_OPERATION);
-	}
-}
-
-void Hosts::cancelTransaction() {
-	if (!db.conn
-			|| sqlite3_exec(db.conn, "ROLLBACK", nullptr, nullptr, nullptr)
-					!= SQLITE_OK) {
-		throw Exception(EX_OPERATION);
-	}
+	SQLite::finalize(stmt.pi);
+	SQLite::finalize(stmt.pq);
+	SQLite::finalize(stmt.pd);
+	SQLite::finalize(stmt.pl);
+	stmt = { nullptr, nullptr, nullptr, nullptr };
 }
 
 } /* namespace wanhive */